requirements.txt
azure_storage==0.37.0
pandas==2.2.2
pyspark==3.5.1
